[FLINK-11843] Move DispatcherRunnerImplTest#testJobRecoveryUnderLeaderChange to DefaultDispatcherRunnerITCase
The DispatcherRunnerImplTest#testJobRecoveryUnderLeaderChange has been moved to the
DefaultDispatcherRunnerITCase#leaderChange_withBlockingJobManagerTermination_doesNotAffectNewLeader.
