+++ flink-runtime/src/test/java/org/apache/flink/runtime/dispatcher/DispatcherResourceCleanupTest.java
@@ -204,11 +204,11 @@ public class DispatcherResourceCleanupTest extends TestLogger {
 				() -> CompletableFuture.completedFuture(resourceManagerGateway),
 				blobServer,
 				heartbeatServices,
-				UnregisteredMetricGroups.createUnregisteredJobManagerMetricGroup(),
 				archivedExecutionGraphStore,
 				fatalErrorHandler,
 				VoidHistoryServerArchivist.INSTANCE,
 				null,
+				UnregisteredMetricGroups.createUnregisteredJobManagerMetricGroup(),
 				jobGraphWriter,
 				jobManagerRunnerFactory));
